Support single files (not just folders) Read file and / or directory list from dedicated file

+++ src/main/java/org/whitesource/fs/WhitesourceFSAgent.java
@@ -25,6 +25,7 @@ import org.whitesource.agent.api.dispatch.UpdateInventoryResult;
 import org.whitesource.agent.api.model.AgentProjectInfo;
 import org.whitesource.agent.api.model.Coordinates;
 import org.whitesource.agent.api.model.DependencyInfo;
+import org.whitesource.agent.client.ClientConstants;
 import org.whitesource.agent.client.WhitesourceService;
 import org.whitesource.agent.client.WssServiceException;
 import org.whitesource.agent.report.OfflineUpdateRequest;
@@ -37,8 +38,6 @@ import java.text.MessageFormat;
 import java.util.*;
 
 import static org.whitesource.fs.Constants.*;
-
-import org.whitesource.agent.client.ClientConstants;
 
 /**
  * @author Itai Marko
@@ -252,29 +251,40 @@ public class WhitesourceFSAgent {
 
         // scan directories
         int totalFiles = 0;
-        Map<File, String[]> fileMap = new HashMap<File, String[]>();
+        Map<File, Collection<String>> fileMap = new HashMap<File, Collection<String>>();
         for (String scannerBaseDir : scannerBaseDirs) {
-            logger.info("Scanning Directory {} for Matching Files (may take a few minutes)", scannerBaseDir);
-            DirectoryScanner scanner = new DirectoryScanner();
-            scanner.setBasedir(scannerBaseDir);
-            scanner.setIncludes(includes.split(INCLUDES_EXCLUDES_SEPARATOR_REGEX));
-            scanner.setExcludes(excludes.split(INCLUDES_EXCLUDES_SEPARATOR_REGEX));
-            if (StringUtils.isNotBlank(globCaseSensitive)) {
-                if (globCaseSensitive.equalsIgnoreCase("true") || globCaseSensitive.equalsIgnoreCase("y")) {
-                    scanner.setCaseSensitive(true);
-                } else if (globCaseSensitive.equalsIgnoreCase("false") || globCaseSensitive.equalsIgnoreCase("n")) {
-                    scanner.setCaseSensitive(false);
+            File file = new File(scannerBaseDir);
+            if (file.exists()) {
+                if (file.isDirectory()) {
+                    logger.info("Scanning Directory {} for Matching Files (may take a few minutes)", scannerBaseDir);
+                    DirectoryScanner scanner = new DirectoryScanner();
+                    scanner.setBasedir(scannerBaseDir);
+                    scanner.setIncludes(includes.split(INCLUDES_EXCLUDES_SEPARATOR_REGEX));
+                    scanner.setExcludes(excludes.split(INCLUDES_EXCLUDES_SEPARATOR_REGEX));
+                    if (StringUtils.isNotBlank(globCaseSensitive)) {
+                        if (globCaseSensitive.equalsIgnoreCase("true") || globCaseSensitive.equalsIgnoreCase("y")) {
+                            scanner.setCaseSensitive(true);
+                        } else if (globCaseSensitive.equalsIgnoreCase("false") || globCaseSensitive.equalsIgnoreCase("n")) {
+                            scanner.setCaseSensitive(false);
+                        } else {
+                            logger.error("Bad {}. Received {}, required true/false or y/n", Constants.CASE_SENSITIVE_GLOB_PROPERTY_KEY, globCaseSensitive);
+                            scmConnector.deleteCloneDirectory();
+                            System.exit(-1); // TODO this is within a try frame. Throw an exception instead
+                        }
+                    }
+                    scanner.scan();
+                    File basedir = scanner.getBasedir();
+                    String[] fileNames = scanner.getIncludedFiles();
+                    fileMap.put(basedir, Arrays.asList(fileNames));
+                    totalFiles += fileNames.length;
                 } else {
-                    logger.error("Bad {}. Received {}, required true/false or y/n", Constants.CASE_SENSITIVE_GLOB_PROPERTY_KEY, globCaseSensitive);
-                    scmConnector.deleteCloneDirectory();
-                    System.exit(-1); // TODO this is within a try frame. Throw an exception instead
+                    // handle file
+                    fileMap.put(file.getParentFile(), Arrays.asList(file.getName()));
+                    totalFiles++;
                 }
+            } else {
+                logger.info(MessageFormat.format("File {0} doesn't exist", scannerBaseDir));
             }
-            scanner.scan();
-            File basedir = scanner.getBasedir();
-            String[] fileNames = scanner.getIncludedFiles();
-            fileMap.put(basedir, fileNames);
-            totalFiles += fileNames.length;
         }
         logger.info(MessageFormat.format("Total Files Found: {0}", totalFiles));
 
@@ -296,7 +306,7 @@ public class WhitesourceFSAgent {
         List<DependencyInfo> dependencyInfos = new ArrayList<DependencyInfo>();
         displayProgress(0, totalFiles);
         int index = 1;
-        for (Map.Entry<File, String[]> entry : fileMap.entrySet()) {
+        for (Map.Entry<File, Collection<String>> entry : fileMap.entrySet()) {
             for (String fileName : entry.getValue()) {
                 DependencyInfo originalDependencyInfo = factory.createDependencyInfo(entry.getKey(), fileName);
                 if (originalDependencyInfo != null) {
